Exception utilities for JNI code
Make it possible for JNI code to throw new exceptions. For that we'll
have to instantiate exception objects with NewObject and then throw
them with Throw. JNI also provides a ThrowNew method, but it accepts
a message while we would like to pass an error code.

In order to instantiate an object we need to know its class and have
a method ID for a suitable constructors. Class and method lookup are
quite expensive so we cache the values. Note that we have to pin
class references in global memory in order to keep method ID valid.
Otherwise JVM is free to unload the classes (while still keeping
the JNI shared library loaded).

Also note that constructing an exception object may also fail with
another exception (most likely an OutOfMemoryError, if ever). In this
case we let the first exception to win. This is semantics of failing
"throw new Exception(...)" in Java too.