Teach natives to throw: key generation
Instead of simply returning NULL, throw an appropriate exception that
stores the original Themis error code.

JNI exception handling is asynchronous, so we still need to return from
the JNI call. After that the signaled exception will be throw into Java
code.

KeyGenerationException is an unchecked exception, but we add a "throws"
specified to the native method regardless. We still leave the error
checking code in place, just in case "null" value is returned.