Wipe output on failure: Secure Cell
Insert soter_wipe() calls into the failure path to wipe partially
written output than might contain sensitive data.

Make sure that miscellaneous local variables are always initialized
before use. This reduces probability of dumb mistakes and often makes
error handling path easier to understand.

Also, slowly get rid of magic macros like THEMIS_CHECK. Replace them
with explicit error checks in plain C. We lose debuggability a bit,
but DEBUG output to stderr is not used too often now.

Here is the first time we use the following wiping pattern:

    if (res != THEMIS_SUCCESS && res != THEMIS_BUFFER_TOO_SMALL) {
        soter_wipe(output_buffer, *output_buffer_length);
    }

We wipe on any error that is not THEMIS_BUFFER_TOO_SMALL, the reason is
that if the buffer is too small then we have already written the desired
buffer length into the provided output variable and thus it is not safe
to wipe the buffer based on that length. However, on these code paths
we do not write anything in the buffer in the first place so there is no
need to wipe them. Normally we don't even perform this check as measure
code paths exit early with plain "return". This is just a failsafe in
case the internal function returns THEMIS_BUFFER_TOO_SMALL (some do).

Note that we always dereference output_buffer_length for wiping. For
this reason make sure that we check it for NULL on entry to functions.

While we're here, I've also inserted additional soter_wipe() calls
to wipe away various derived keys and IVs that we keep on stack.

This covers Themis code for Secure Cell API:

  - themis_secure_cell_encrypt_seal
  - themis_secure_cell_decrypt_seal
  - themis_secure_cell_encrypt_token_protect
  - themis_secure_cell_decrypt_token_protect
  - themis_secure_cell_encrypt_context_imprint
  - themis_secure_cell_decrypt_context_imprint

Soter internals have not been audited.

Now I'm mostly happy with how secure_cell.c looks, but sym_enc_message.c
leaves... a lot of opportunities for improvements in code readability
and clarity.