Wipe output on failure: key generation
The core of this change is soter_wipe() call on failure. Current code
of key generation error handling is not really readable so we refactor
it a bit. Common error handling is extracted into a function and the
code is streamlined now.

Symmetric key generation simply calls soter_rand() which wipes after
itself as an adult. We leave a comment just in case the code gets more
complex in the future, so that the developers don't forget to add a
soter_wipe() call if necessary.

This covers Themis code for key generation API:

  - themis_gen_sym_key
  - themis_gen_rsa_key_pair
  - themis_gen_ec_key_pair

Soter internals have not been audited.