Drop openssl.framework from the project
Remove CLOpenSSL from Carthage dependencies, remove all remaining
references to this framework, stop embedding it into tests. Now Themis
is using embedded BoringSSL when building from Carthage.

Since CLOpenSSL is a static framework, it does not need to linked into
the application (since Themis 0.13). Removing it as a dependency is not
a breaking changea as the end-user applications don't use it directly.