Bring back "undefined" into TypeError tests
Previous change has removed it from the list of 'invalid values' because
in TypeScript "undefined" is interpreted as "omitted argument" which is
actually allowed in some places and does not cause failures.

Instead of removing "undefined", skip it for those APIs that accept
omitted arguments, but still test other APIs with "undefined".