(Re)enable dead code elimination
Newer release of Emscripten finally fixed an issue in its dead code
elimiation pass that treated the entire WasmThemis as dead code.

Drop LINKABLE=1 flag and replace it MAIN_MODULE=2, instructing wasm-ld
to build a main module *and* perform dead code elimination. It takes
considerably more time, but results in 10 KB of WASM that we actally use
instead of the current 1.6 MB of BoringSSL stuff that we don't use.