Compatibility "data" properties for keys
Currently in TypeScript code SymmetricKey, PublicKey, PrivateKey are not
inherited from Uint8Array. Instead, they provide a "data" field holding
the array.

We're going to do proper inheritance later, but for now, make the code
bug-for-bug compatible.

"this.data" is intentionally made a *copy* of "this", not a
self-reference, because self-referencing data breaks some assertions.